pyproject.toml
[project]
name = "mahirosearch"
version = "0.2.0"
description = "AI semantic search for local files - Pure PySide6 Desktop Application"
requires-python = ">=3.11"
dependencies = [
    "PySide6>=6.6.0",
    "PySide6-Fluent-Widgets>=1.11.1",
    "pydantic-settings>=2.2.0",
    "pyyaml>=6.0",
    "lancedb>=0.6.0",
    "pyarrow>=15.0",
    "httpx>=0.27.0",
    "tiktoken>=0.6.0",
    "pdfplumber>=0.10.0",
    "python-docx>=1.1.0",
    "watchdog>=4.0.0",
]

[project.scripts]
mahirosearch = "main:main"

[tool.setuptools]
py-modules = []
